requirements.txt
alive-progress==1.6.2
amqp==2.6.1
appnope==0.1.2
APScheduler==3.6.3
argon2-cffi==20.1.0
asgiref==3.3.1
async-generator==1.10
attrs==20.3.0
Automat==20.2.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
bleach==3.2.1
blinker==1.4
Brotli==1.0.9
cacheout==0.11.2
celery==4.4.6
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
Click==7.0
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.4
commonmark==0.9.1
constantly==15.1.0
cryptography==3.2.1
curlify==2.2.1
decorator==4.4.2
defusedxml==0.6.0
diagrams==0.17.0
entrypoints==0.3
fake-useragent==0.1.11
Faker==4.14.0
filter==0.0.0.20200724
Flask==1.1.2
Flask-Limiter==1.4
freezegun==1.1.0
future==0.18.2
gevent==20.12.0
graphviz==0.13.2
greenlet==0.4.17
h11==0.11.0
h2==4.0.0
hpack==4.0.0
huey==2.3.0
hyperframe==6.0.0
hyperlink==20.0.1
idna==2.10
importlib-metadata==2.0.0
incremental==17.5.0
ipykernel==5.4.2
ipython==5.7.0
ipython-genutils==0.2.0
ipywidgets==7.6.2
itsdangerous==1.1.0
Jinja2==2.11.2
jsonpath==0.82
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kafka==1.3.5
kaitaistruct==0.9
kombu==4.6.11
langid==1.1.6
ldap3==2.8.1
limits==1.5.1
MarkupSafe==1.1.1
marshmallow==3.10.0
meteo-downloader==1.0
mistune==0.8.4
mitmproxy==5.3.0
msgpack==1.0.0
msgpack-python==0.5.6
munch==2.5.0
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.1.6
numpy==1.19.2
packaging==20.8
paho-mqtt==1.5.1
pandas==1.1.3
pandocfilters==1.4.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
pika==1.1.0
Pillow==8.0.1
pretty-errors==1.2.19
prometheus-client==0.9.0
prompt-toolkit==3.0.8
protobuf==3.13.0
ptyprocess==0.6.0
publicsuffix2==2.20191221
pyasn1==0.4.8
pycparser==2.20
Pygments==2.7.3
PyHamcrest==2.0.2
pylane==0.0.9
pymongo==3.11.0
pynsq==0.9.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.1
pyrsistent==0.17.3
PySnooper==0.4.2
PySocks==1.7.1
pytesseract==0.3.7
python-dateutil==2.8.1
python-telegram-bot==13.1
pytz==2020.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
redis==3.0.1
redis-py-cluster==2.0.0
requests==2.24.0
rich==9.6.1
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
selenium==3.141.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.15.0
socks==0
sortedcontainers==2.2.2
soupsieve==2.0.1
SQLAlchemy==1.3.19
tenacity==6.3.1
terminado==0.9.1
testpath==0.4.4
text-unidecode==1.3
tornado==6.0.4
tqdm==4.51.0
traitlets==5.0.5
Twisted==20.3.0
typing-extensions==3.7.4.3
tzlocal==2.1
urllib3==1.25.11
urwid==2.1.2
verifier==1.0.0
vine==1.3.0
wcwidth==0.2.5
webencodings==0.5.1
websocket==0.2.1
websocket-client-py3==0.15.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
ws4py==0.5.1
wsproto==0.15.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.2.0
zstandard==0.14.0
